requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib==3.8.0
seaborn==0.13.0
scikit-learn==1.3.0
SQLAlchemy==2.0.21
pytest==7.4.2
black==23.9.1
flake8==6.1.0 